chore: use string instead of enum
